requirements.txt
asgiref==3.7.2
Django==4.2.17
django-extensions==4.1
djangorestframework==3.16.1
dnspython==2.8.0
pymongo==4.15.5
ruff==0.14.10
shelve_cache==0.0.4
sqlparse==0.5.5
